requirements.txt
Django==3.1
mysqlclient==2.0.3
django-mysql==3.10.0
django-environ==0.4.4
Pillow==8.1.0
django-taggit=1.3.0
